Day 03 - Extremely unclean, overengineered Version
Part 1 has a solve that we don't know where it comes from. It works, but
it's probably only good to work, not to be pretty, useful, extensible or
anything.

Part 2 has a useful solve (and actually solves part 1 by changing
`digit_count`), but took way too long to not try to over engineer. It's
a long Tuesday.

Cleanup to come.