Day08 - Containers are Hell
It is unwise to have containers of containers. Both parts compute all
euclidean distances, and sort them, and use that as basis.

Part 1 takes the first `conn_count`, and starts creating circuits. The
idea is to start with sets of two elements, iterate through all sets,
find indices in other sets, and combine them together. Probably a set
here was a bad choice of container, could've just assigned circuit
indices or something.

Part 2 actually tries to be smart. We keep a note of seen indices. We
add only new boxes to our circuit. Once we've connected (seen) every
index, we return the last two connected.