Day 07 fix
Rather hard to debug issue with `std::accumulate`. The return type is
the type of `init` (third argument). The given `0` defaults to `int`. It
gets automatically cast to `size_t` in the lambda, but gets converted
back into `int` on return of `std::accumulate`, which overflows when
stored into `timelines`. The compiler, even with `pedantic` and `all`,
doesn't capture this.

The solution to detect the narrowing conversion is to explicitly build a
`size_t`, or to otherwise fix the problem, declare the `init` value as
`0ull`.